Add documentation and reorganize dashboard and metadata packages
