+++ misk/src/main/kotlin/misk/web/dashboard/ValidWebEntry.kt
@@ -1,4 +1,4 @@
-package misk.web
+package misk.web.dashboard
 
 import com.google.common.base.CharMatcher
