+++ misk/src/main/kotlin/misk/web/metadata/ServiceMetadataAction.kt
@@ -1,4 +1,4 @@
-package misk.web.actions
+package misk.web.metadata
 
 import misk.config.AppName
 import misk.environment.Environment
@@ -6,6 +6,7 @@ import misk.security.authz.Unauthenticated
 import misk.web.Get
 import misk.web.RequestContentType
 import misk.web.ResponseContentType
+import misk.web.actions.WebAction
 import misk.web.mediatype.MediaTypes
 import javax.inject.Inject
 import javax.inject.Singleton
@@ -22,7 +23,8 @@ class ServiceMetadataAction @Inject constructor(
   @ResponseContentType(MediaTypes.APPLICATION_JSON)
   @Unauthenticated
   fun getAll(): Response {
-    return Response(serviceMetadata = optionalBinder.serviceMetadata)
+    return Response(
+      serviceMetadata = optionalBinder.serviceMetadata)
   }
 
   data class Response(val serviceMetadata: ServiceMetadata)
@@ -34,7 +36,8 @@ class ServiceMetadataAction @Inject constructor(
 @Singleton
 class OptionalBinder @Inject constructor(@AppName val appName: String) {
   @com.google.inject.Inject(optional = true)
-  var serviceMetadata: ServiceMetadata = ServiceMetadata(appName, Environment.fromEnvironmentVariable())
+  var serviceMetadata: ServiceMetadata =
+    ServiceMetadata(appName, Environment.fromEnvironmentVariable())
 }
 
 data class ServiceMetadata(
